EventSystem
I recoded the old waterdrop we had and implemented a wip eventmanager.
Waterdrop technically works!
Howver I have yet to make the messages configurable aswell as some fields in the config straight up doesnt do anything anymore.
Permission nodes are pretty much not implemented as of now but will be available when the alpha releases.
Implemented #2