One in the chamber alpha ready
One in the chamber is now playable and presentable for an alpha stage.
There are some stuff I completely want to changeas im not happy with it yet. I will continue to work with that after Alpha 1.  Closed #4  Also removed WorldGuard as a dependency as we do not need it because we have our own region system.